+++ include/gaussianelimination.hh
@@ -2,7 +2,7 @@
 #define GAUSSIAN_ELIMINATION_GUARD
 
 
-void solve_simultaneous(float* X, float** A, float* R, int size);
-void invert_matrix(float** A, float **R, int size);
-void predivide(float** A, float **B, int size, int cols);
+void solve_simultaneous(double* X, double** A, double* R, int size);
+void invert_matrix(double** A, double **R, int size);
+void predivide(double** A, double **B, int size, int cols);
 #endif//GAUSSIAN_ELIMINATION_GUARD
